+++ blogit/sitemaps.py
@@ -4,14 +4,15 @@ from __future__ import unicode_literals
 from django.contrib.sitemaps import Sitemap
 
 from blogit.models import Post
+from blogit import settings as bs
 
 
 class BlogitSitemap(Sitemap):
-    priority = 0.5
-    changefreq = 'weekly'
+    priority = bs.SITEMAP_PRIORITY
+    changefreq = bs.SITEMAP_CHANGEFREQ
 
     def items(self):
-        return Post.objects.translated().published()
+        return Post.objects.public().published()
 
     def lastmod(self, obj):
         return obj.last_modified